# swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This package provides a REST API to the Ignition Fuel server.",
    "title": "Ignition Fuel Server REST API",
    "contact": {
      "email": "info@openrobotics.org"
    },
    "license": {
      "name": "Apache 2.0"
    },
    "version": "0.1.0"
  },
  "host": "fuel.ignitionrobotics.org",
  "basePath": "/1.0",
  "paths": {
...
    "/organizations": {
      "get": {
        "description": "Get the list of organizations. Organizations will be returned paginated,\nwith pages of 20 organizations by default. The user can request a\ndifferent page with query parameter 'page' (first page is value 1).\nThe page size can be controlled with query parameter 'per_page',\nwith a maximum of 100 items per page.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "organizations"
        ],
        "summary": "List organizations",
        "operationId": "listOrganizations",
        "responses": {
          "200": {
            "description": "OrganizationResponses",
            "schema": {
              "$ref": "#/definitions/OrganizationResponses"
            }
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      },
      "post": {
        "description": "Creates a new organization. Note: the user identity will be retrieved from the passed JWT.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "organizations"
        ],
        "summary": "Create organization",
        "operationId": "createOrganization",
        "responses": {
          "200": {
            "description": "OrganizationResponse",
            "schema": {
              "$ref": "#/definitions/OrganizationResponse"
            }
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      }
    },
    "/organizations/{name}": {
      "get": {
        "description": "Return an organization given its name and a valid JWT.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "organizations"
        ],
        "summary": "Get an organization",
        "operationId": "singleOrganization",
        "responses": {
          "200": {
            "description": "OrganizationResponse",
            "schema": {
              "$ref": "#/definitions/OrganizationResponse"
            }
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      },
      "delete": {
        "description": "Deletes an organization given its name and a valid JWT.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "organizations"
        ],
        "summary": "Delete an organization",
        "operationId": "deleteOrganizations",
        "responses": {
          "200": {
            "description": "OrganizationResponse",
            "schema": {
              "$ref": "#/definitions/OrganizationResponse"
            }
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      },
      "patch": {
        "description": "Update an organization",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "https"
        ],
        "tags": [
          "organizations"
        ],
        "summary": "Update an organization",
        "operationId": "organizationUpdate",
        "responses": {
          "200": {
            "description": "OrganizationResponse",
            "schema": {
              "$ref": "#/definitions/OrganizationResponse"
            }
          },
          "default": {
            "$ref": "#/responses/fuelError"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrMsg": {
      "description": "ErrMsg is serialized as JSON, and returned if the request does not succeed\nTODO: consider making ErrMsg an 'error'",
      "type": "object",
      "properties": {
        "errcode": {
          "description": "Internal error code.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrCode"
        },
        "errid": {
          "description": "Generated ID for easy tracking in server logs",
          "type": "string",
          "x-go-name": "ErrID"
        },
        "extra": {
          "description": "Extra information/arguments associated to Error message.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Extra"
        },
        "msg": {
          "description": "Error message.",
          "type": "string",
          "x-go-name": "Msg"
        },
        "remote-address": {
          "description": "Associated request remote address, if applicable",
          "type": "string",
          "x-go-name": "RemoteAddress"
        },
        "route": {
          "description": "Associated request Route, if applicable",
          "type": "string",
          "x-go-name": "Route"
        },
        "user-agent": {
          "description": "Associated request User-Agent, if applicable",
          "type": "string",
          "x-go-name": "UserAgent"
        }
      },
      "x-go-package": "gitlab.com/ignitionrobotics/web/ign-go"
    },
...
    "OrganizationResponse": {
      "type": "object",
      "title": "OrganizationResponse stores organization information used in REST responses.",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "private": {
          "type": "boolean",
          "x-go-name": "Private"
        }
      },
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/users"
    },
    "OrganizationResponses": {
      "description": "OrganizationResponses is a slice of OrganizationResponse",
      "type": "array",
      "items": {
        "$ref": "#/definitions/OrganizationResponse"
      },
      "x-go-package": "github.com/gazebo-web/fuel-server/bundles/users"
    },
...
  },
  "responses": {
...
    "fuelError": {
      "description": "Ign Fuel error serialized as JSON",
      "schema": {
        "$ref": "#/definitions/ErrMsg"
      }
    },
...
  }
}